+++ lib/factory/action_factory.dart
@@ -12,7 +12,6 @@ import '../bloc/editor_bloc.dart';
 import '../factory/widget_factory.dart';
 import '../model/action_context.dart';
 import '../model/node_spec.dart';
-import '../model/solution.dart';
 import '../model/solution_exception.dart';
 import '../util/extensions.dart';
 import '../util/parser.dart';
@@ -132,8 +131,10 @@ class ActionFactory {
       return;
     }
 
+    final runState = actionContext ?? {};
     var currentValue = eventValue;
     NodeSpec? currentAction = action;
+
     while (currentAction != null) {
       if (!buildContext.mounted) {
         logError(
@@ -142,7 +143,7 @@ class ActionFactory {
       }
 
       final context =
-          ActionContext(state, actionContext ?? {}, currentValue, buildContext);
+          ActionContext(state, runState, currentValue, buildContext);
       if (!await preExecute(currentAction, context)) {
         break;
       }
@@ -240,7 +241,7 @@ class ActionFactory {
       NodeSpec action, ActionResult result, ActionContext context) async {
     if (result.returnData != null || action.props["returnName"] != null) {
       final key = action.props["returnName"] ?? "value";
-      context.state[key] = result.returnData;
+      context.actionContext[key] = result.returnData;
     }
 
     // if (result.emitStates != null) {
@@ -262,37 +263,8 @@ class ActionFactory {
 
   /// Returns a Map with the evaluation context used when sanitizing properties of a [NodeSpec].
   /// Used to resolve properties that use placeholders as values, like "${state.firstName}" or "${env.api_uri}".
-  Map getEvaluatorContext(Object? value, Map state, Map? specContext) {
-    final mediaQueryData = MediaQueryData.fromView(
-        WidgetsBinding.instance.platformDispatcher.views.single);
-
-    final map = {};
-    if (specContext != null) map.addAll(specContext);
-    map.addAll({
-      "null": null,
-      "languages": Solution.languages,
-      "language": Solution.language,
-      "env": Solution.environmentVariables,
-      "global": Lowder.globalVariables,
-      "state": state,
-      "value": value,
-      "media": {
-        "isWeb": kIsWeb,
-        "isMobile": !kIsWeb && mediaQueryData.size.shortestSide < 600,
-        "isTablet": !kIsWeb && mediaQueryData.size.shortestSide >= 600,
-        "isAndroid": kIsWeb ? false : Platform.isAndroid,
-        "isIOS": kIsWeb ? false : Platform.isIOS,
-        "isWindows": kIsWeb ? false : Platform.isWindows,
-        "isMacOS": kIsWeb ? false : Platform.isMacOS,
-        "isLinux": kIsWeb ? false : Platform.isLinux,
-        "isFuchsia": kIsWeb ? false : Platform.isFuchsia,
-        "portrait": mediaQueryData.orientation == Orientation.portrait,
-        "landscape": mediaQueryData.orientation == Orientation.landscape,
-        // "version": Platform.version,
-      }
-    });
-    return map;
-  }
+  Map getEvaluatorContext(Object? value, Map state, Map? specContext) =>
+      Lowder.properties.getEvaluatorContext(value, state, specContext);
 
   /// Exception handling when executing an [action]
   Future<RetryAction> handleException(