+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.4
+
+* Fixed an issue where an Action's value wasn't available on the next Actions.
+* Updated example.
+* Some cosmetic work and cleanup.
+
 ## 0.1.3
 
 * Yet more file formatting to respect pub.dev standards.